requirements.txt
pillow
fitz